Move chaching cache system to datastore
