+++ packages/libs/configuration/src/index.ts
@@ -1,6 +1,5 @@
 export * from "./autorest-configuration";
 export * from "./autorest-raw-configuration";
-export * from "./caching-file-system";
 export * from "./configuration-manager";
 export * from "./configuration-file-resolver";
 export * from "./configuration-loader";
