+++ packages/libs/datastore/src/file-system/index.ts
@@ -1,4 +1,5 @@
+export * from "./caching-file-system";
+export * from "./enhanced-file-system";
 export * from "./file-system";
 export * from "./memory-file-system";
 export * from "./real-filesystem";
-export * from "./enhanced-filesystem";
